+++ Examples/Color With String/ColorUtilsExample/ViewController.m
@@ -15,10 +15,14 @@
 {
     [super viewDidLoad];
 	
+    //register a custom color
+    UIImage *plaid = [UIImage imageNamed:@"Plaid.jpg"];
+    [UIColor registerColor:[UIColor colorWithPatternImage:plaid] forName:@"plaid"];
+    
     //find all labels in view and style them using their text as as a color value
     for (UIView *view in self.view.subviews)
     {
-        if ([view isKindOfClass:[UILabel class]])
+        if ([view isKindOfClass:[UILabel class]] && view.tag == 0)
         {
             UILabel *label = (UILabel *)view;
             UIColor *color = [UIColor colorWithString:label.text];